Add GisTile and GisWms and update wms_test accordingly
GisTile:
 * Represents a lat-lon tile
 * May contain pointers to child tiles of higher resolution

GisWms:
 * Contains Web Map Service support functions

GisWms and GisTile roughly translate to WmsInfo and WmsCacheNode. The main
difference being that the new versions are more independent of each other so
non-wms downloaders can be used in the future.

The other major difference is that GisTile is not transparent as WmsCacheNode
was. That is, there is no gis_tile_fetch function which automatically updates
the tiles.