Add Shuttle Radar Topography Mission plugin and height code
For the most part, the SRTM plugin is simmilar to the BMNG plugin. The
loaders are more complicated because it generates both a bil (unsigned
short array) and a pixmap from the downloaded files.

Highmaps are in two parts when set_height_func is called with
update=TRUE, the heights of all the vertices for a given tile are
updated by calling height_func and using the returned value as the new
vertex height.

In addition to updating the vertex height immediately, the height
function is also stored in each vertex and used later during the
split-merge process.

The user_data passed with set_height_func could theoretically be used to
speed searching for the correct tile during the split-merge process, but
this is currently not done. (Each call to the srtm height function does
a search for the correct tile)