Convert from xyz to lle as the main roam coordinates
This keeps splits along latitude lines instead of along great circles.
It is a little less accurate, but makes texture mapping much simpler, at
least with Mercator projections.

If we ever manage to get lle->xyz conversions done in an OpenGL shader,
this should make the transition a little easier as well.

     Old Way           New Way
    .-''/'''-.        .-''/'''-.
   / __/_____ \      /.__/____..\
  /''  |     ''\    /    |       \
  *..__*_____..*    *..__*_____..*
  \    |       /    \    |       /
   \   \      /      \   \      /
    '-..\...-'        '-..\...-'