Give extra weight to "edge" triangles when splitting
"Edge" triangles are triangles where one of the edges of the triangle
makes up an edge off the mesh. For example, the triangle E below is an
edge triangle due to the top edge.

This helps prevent jagged "edges" on the tops of mountains, etc. It also
helps round out the edges of the globe  when viewed from far away.

  ________________
  \    ______    /
   \  /\ E  /\  /
    \/__\  /__\/
         \/