Add mouse enter/leave signals to objects
This uses the OpenGL Selection render mode to determine which objects
the mouse is over. It requires fairly tight integration between
GritsOpenGL and GritsObject.

The signal code is handled internally by GritsObject. However, most of
the actual selection matching work is done by GritsOpenGL.

Object types that wish to improve performance can implement the pick()
function in addition to the draw() function. The pick function is used
when performing selection matching. It works similar to draw, but does
not need to do textures/lighting/opacity/etc.