Improve performance of GritsPoly
Use separate display lists, and only create them and perform
tessellation when they are actually used.

This is especially useful for large number of polygons which aren't
always displayed (or may never be displayed).

Generating the display lists from the first draw() also avoids a large
number of idle callbacks.