Add GritsLine
GritsLine is similar to GritsPoly, but does not draw a filled polygon.
As such it does not need to do any tessellation.

Currently it draws the points directly instead of using a display list.
This should make it easier to updated the line at runtime without
needing to create a new object.